Restructure Scripts
