+++ scripts/rate-scraper/scrape.ts
@@ -3,13 +3,7 @@
 import { readFile, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import type { RatesFile } from "@/lib/schemas/rate";
-import {
-	buildHistoryFromResults,
-	previewHistory,
-	saveHistoryFile,
-} from "./history/build-history";
 import { appendChangeset, ensureHistoryExists } from "./history/changeset";
-import { isHistoricalProvider, scrapeHistorical } from "./history/historical";
 import { aibProvider } from "./providers/aib";
 import { avantProvider } from "./providers/avant";
 import { boiProvider } from "./providers/boi";
@@ -194,156 +188,11 @@ function formatCommitDescription(results: LenderResult[]): string {
 	return lines.join("\n");
 }
 
-async function runHistoricalScrape(lenderId: string, args: string[]) {
-	const provider = providers[lenderId];
-
-	if (!provider) {
-		console.error(`Unknown lender: ${lenderId}`);
-		console.error(`Available lenders: ${Object.keys(providers).join(", ")}`);
-		process.exit(1);
-	}
-
-	if (!isHistoricalProvider(provider)) {
-		console.error(`Provider ${lenderId} does not support historical scraping`);
-		console.error("Historical providers must implement parseHtml() method");
-		process.exit(1);
-	}
-
-	// Parse options
-	const dryRun = args.includes("--dry-run");
-	const buildHistory = args.includes("--build");
-	const mergeWithExisting = args.includes("--merge");
-	const fromArg = args.find((a) => a.startsWith("--from="));
-	const toArg = args.find((a) => a.startsWith("--to="));
-	const maxArg = args.find((a) => a.startsWith("--max="));
-
-	const fromDate = fromArg?.split("=")[1];
-	const toDate = toArg?.split("=")[1];
-	const maxSnapshots = maxArg
-		? Number.parseInt(maxArg.split("=")[1], 10)
-		: undefined;
-
-	console.log(`\n${"=".repeat(60)}`);
-	console.log(`Historical scrape for ${provider.name} (${lenderId})`);
-	console.log("=".repeat(60));
-	console.log(`URL: ${provider.url}`);
-	if (fromDate) console.log(`From: ${fromDate}`);
-	if (toDate) console.log(`To: ${toDate}`);
-	if (maxSnapshots) console.log(`Max snapshots: ${maxSnapshots}`);
-	if (dryRun) console.log("Mode: DRY RUN (no fetching)");
-	if (buildHistory)
-		console.log(
-			"Mode: BUILD HISTORY" +
-				(mergeWithExisting ? " (merge with existing)" : ""),
-		);
-	console.log("");
-
-	const report = await scrapeHistorical(provider, {
-		fromDate,
-		toDate,
-		maxSnapshots,
-		dryRun,
-	});
-
-	if (!dryRun && report.results.length > 0) {
-		console.log(`\n${"=".repeat(60)}`);
-		console.log("RESULTS SUMMARY");
-		console.log("=".repeat(60));
-		for (const result of report.results) {
-			console.log(
-				`  ${result.timestamp}: ${result.rates.length} rates (hash: ${result.hash.slice(0, 8)}...)`,
-			);
-		}
-
-		// Build history if requested
-		if (buildHistory) {
-			console.log(`\n${"=".repeat(60)}`);
-			console.log("BUILDING HISTORY FILE");
-			console.log("=".repeat(60));
-
-			const { history, report: buildReport } = await buildHistoryFromResults(
-				lenderId,
-				report.results,
-				{
-					mergeWithExisting,
-					validateAgainstCurrent: true,
-				},
-			);
-
-			previewHistory(history);
-
-			console.log("\nBuild Report:");
-			console.log(`  Baseline: ${buildReport.baselineTimestamp}`);
-			console.log(`  Baseline rates: ${buildReport.baselineRatesCount}`);
-			console.log(`  Changesets: ${buildReport.changesetsCount}`);
-			console.log(`  Final hash: ${buildReport.finalHash.slice(0, 12)}...`);
-
-			if (buildReport.currentRatesHash) {
-				console.log(
-					`  Current rates hash: ${buildReport.currentRatesHash.slice(0, 12)}...`,
-				);
-				if (buildReport.hashesMatch) {
-					console.log(
-						"  ✓ Hashes match - history is consistent with current rates",
-					);
-				} else {
-					console.log(
-						"  ⚠ Hashes DO NOT match - there may be a gap in history",
-					);
-				}
-			}
-
-			// Save the history file
-			await saveHistoryFile(history);
-			console.log(
-				`\n✓ History file saved to data/rates/history/${lenderId}.json`,
-			);
-		}
-	}
-
-	if (report.errors.length > 0) {
-		console.log("\nERRORS:");
-		for (const error of report.errors) {
-			console.log(`  ${error.timestamp}: ${error.error}`);
-		}
-	}
-}
-
 async function main() {
 	const args = process.argv.slice(2);
 	const hasAll = args.includes("--all");
-	const hasHistorical = args.includes("--historical");
 	const writeUpdates = args.includes("--write-updates");
 	const lenderId = args.find((a) => !a.startsWith("--"));
-
-	// Historical scraping mode
-	if (hasHistorical) {
-		if (!lenderId) {
-			console.error(
-				"Usage: bun run rates:scrape --historical <lender-id> [options]",
-			);
-			console.error("Options:");
-			console.error("  --from=YYYY-MM-DD  Start date for snapshots");
-			console.error("  --to=YYYY-MM-DD    End date for snapshots");
-			console.error("  --max=N            Maximum number of snapshots");
-			console.error("  --dry-run          List snapshots without fetching");
-			console.error(
-				"  --build            Build and save history file from results",
-			);
-			console.error(
-				"  --merge            Merge with existing history (use with --build)",
-			);
-			console.error(
-				`\nHistorical-enabled lenders: ${Object.entries(providers)
-					.filter(([_, p]) => isHistoricalProvider(p))
-					.map(([id]) => id)
-					.join(", ")}`,
-			);
-			process.exit(1);
-		}
-		await runHistoricalScrape(lenderId, args);
-		return;
-	}
 
 	if (hasAll && lenderId) {
 		console.error("Error: Cannot use --all with a specific lender");
