+++ scripts/rate-scraper/validate.ts
